requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0